bug 1247690 - Move signal handlers to own module
Move signal handlers from models.py to new signals.py.  This requires
using the AppConfig added in Django 1.7.
