+++ webplatformcompat/__init__.py
@@ -1,2 +1,3 @@
 """The BrowserCompat API app."""
 __version__ = '0.1.0'
+default_app_config = 'webplatformcompat.apps.WebPlatformCompatConfig'
